+++ packages/primeng/schematics/ng-add/schema.ts
@@ -1,0 +1,4 @@
+export interface NgAddSchema {
+  theme?: 'aura' | 'material' | 'lara' | 'nora';
+  skipInstall?: boolean;
+}